More complete and realistic performance test, should also compile and run on linux
here's the perf test results with different object allocations
  on winxpsp2, msvc6 debug compile, Pentium 4 3ghz CPU 1Gig RAM
  
It show that search time in neglectable compared to alloc/dealloc time
making me think that if the node alloc is replaced by internal widget member
we can be even faster.

Real Alloc'd objects matching the Widget size have been used this time
for more realistic pointer hash perf test.

Don't know if we would even use that one day but it was quite 
interesting to  have a precise idea of what it could cost (almost nothing)
to add this kind of feature

-------> Results Begin
Number of Widgets =  10000
Allocation time 0.047 seconds
Find time   0 secs,   0 microsec per object
is_valid time   0 secs,   0 microsec per object
DeAllocation time 0.141 secs

Number of Widgets =  100000
Allocation time 0.187 seconds
Find time 0.031 secs, 0.31 microsec per object
is_valid time 0.047 secs, 0.47 microsec per object
DeAllocation time 2.547 secs

Number of Widgets =  250000
Allocation time 0.468 seconds
Find time 0.204 secs, 0.816 microsec per object
is_valid time 0.218 secs, 0.872 microsec per object
DeAllocation time 3.782 secs

Number of Widgets =  500000
Allocation time 0.938 seconds
Find time 0.875 secs, 1.75 microsec per object
is_valid time 0.906 secs, 1.812 microsec per object
DeAllocation time 8.438 secs
<---------- Results End




git-svn-id: http://seriss.com/public/fltk/fltk/branches/fcupd1@5077 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ hash/fltk/lut.h
@@ -13,6 +13,7 @@
 ////////////////////////////////////////////////////////////////////////////////////	
 #include "FL_API.h"
 #include <fltk/types.h>
+#include <stddef.h>
 
 namespace fltk {
 
@@ -59,7 +60,7 @@ namespace fltk {
 	    //! initialize the class internals
 	    static void init();
 	    //! free resources
-	    static void done() {for (size_t i=0; i<hashTableLen; i++) Node::clear(slot[i]);}
+	    static void done() {for (int i=0; i<hashTableLen; i++) Node::clear(slot[i]);}
 	    
 	    //! for statistics purpose count the nuumber of stored nodes in the hashTable
 	    static size_t count();
